Separated original project into five projects:
osuCelebrity-parent: parent project
osuCelebrity-model: interfaces
osuCelebrity-osu: components which communicate with osu
osuCelebrity-twitch: components which communicate with twitch
osuCelebrity-core: components which implement our idea
osuCelebrity: deployment

Note that there are plenty of compile errors, because details need to be
worked out.

Findbugs and checkstyle (google style) are enabled in test phase (test
phase for checkstyle is easier to invoke than verify). Style files for
eclipse and IDEA can be downloaded here:
https://github.com/google/styleguide

Note that there are a couple of bugs and checkstyle violations, which
need to be sorted out after everything is put back together.