Added a robot, which looks for download buttons and clicks them.
The images in osuCelebrity/ need to be in the working directory when the
bot runs. These are client-resolution-specific examples and you might
have to replace them.

If these images cannot be found, the bot will issue a warning and image
recognition will not run. The bot will work just fine otherwise.