Removed handling of osu IRC ERR_NOSUCHNICK.
If a player does not allow messages from people who are not in their
friend list, a ERR_NOSUCHNICK is sent. This was preventing those players
from being spectated (the previous assumption was that they were
offline). We are using BanchoBot !stat now for the same purpose, so we
don't need this mechanism anymore.