New auto-queue mechanism. Closes #52.
The old auto-queue mechanism is still in place, but should be invoked
rarely.

Note that when sorting the queue, the minimum votes are now set to 0.5
to let non-auto-queued players pass auto-queued players, but to not have
them jump to the top immediately. This value might have to be adjusted.

Furthermore, auto-queued players are ignored when calculating the
current queue size when redetermining the remaining time for the current
player.